fix: export runAllFeatureTicks aggregator
