+++ index.html
@@ -1219,7 +1219,7 @@
     <svg id="astralTreeSvg"></svg>
   </div>
 
-  <script type="module" src="ui/index.js"></script>
+  <script type="module" src="src/ui/app.js"></script>
   <script type="module">
     import { mountDevQuickMenu } from "./src/ui/dev/devQuickMenu.js";
     if (document.readyState === "loading") {
